time: int->long
Former-commit-id: 15c78e3089621bfe9edc37be7c6483fb45dbf2e4

+++ back-end-projects/Explorer/src/main/java/com/github/ontio/controller/BlockController.java
@@ -140,7 +140,7 @@ public class BlockController {
      */
     @RequestMapping(value = "/blockCountInTwoWeeks/{time}", method = RequestMethod.GET)
     @ResponseBody
-    public Result blockCountInTwoWeeks(@PathVariable("time") int time) {
+    public Result blockCountInTwoWeeks(@PathVariable("time") long time) {
 
         logger.info("########{}.{} begin...", CLASS_NAME, Helper.currentMethod());
         logger.info("time:{}", time);
